comments and test fixup
